Moving botactions to type folder
